Cleanup Envoy ztunnel implementation and code
