+++ istio.deps
@@ -11,6 +11,6 @@
     "name": "ZTUNNEL_REPO_SHA",
     "repoName": "ztunnel",
     "file": "",
-    "lastStableSHA": "e48b47e5f7e21698ea95a88d04b4faa4302dbe92"
+    "lastStableSHA": "8ed50c9848694801b7eb6ff0b76061b694912041"
   }
 ]
