+++ pkg/test/framework/resource/flags.go
@@ -177,9 +177,6 @@ func init() {
 	flag.BoolVar(&settingsFromCommandLine.Ambient, "istio.test.ambient", settingsFromCommandLine.Ambient,
 		"Indicate the use of ambient mesh.")
 
-	flag.BoolVar(&settingsFromCommandLine.AmbientNativeZtunnel, "istio.test.ambientNativeZtunnel", settingsFromCommandLine.AmbientNativeZtunnel,
-		"Indicate the use of ambient mesh with the native ztunnel implementation.")
-
 	flag.BoolVar(&settingsFromCommandLine.Compatibility, "istio.test.compatibility", settingsFromCommandLine.Compatibility,
 		"Transparently deploy echo instances pointing to each revision set in `Revisions`")
