+++ pkg/test/framework/resource/settings.go
@@ -140,8 +140,6 @@ type Settings struct {
 
 	// Ambient mesh is being used
 	Ambient bool
-	// AmbientNativeZtunnel means ambient is used with https://github.com/istio/ztunnel
-	AmbientNativeZtunnel bool
 
 	// Compatibility determines whether we should transparently deploy echo workloads attached to each revision
 	// specified in `Revisions` when creating echo instances. Used primarily for compatibility testing between revisions
